backend/requirements.txt
fastapi
uvicorn
python-dotenv
httpx
requests
aiofiles
pillow
python-multipart